requirements.txt
# youtube
pandas==2.2.2
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6
odfpy==1.4.1
pyyaml==6.0.1
